// src/swagger/transform/models/sync-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * REST API SPEC FOR MOVESONG FRONTEND
 * REST API SPEC FOR MOVESONG FRONTEND
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface SyncDto
 */
export interface SyncDto {
    /**
     * 
     * @type {number}
     * @memberof SyncDto
     */
    id?: number;
    /**
     * 
     * @type {boolean}
     * @memberof SyncDto
     */
    enabled?: boolean;
    /**
     * 
     * @type {string}
     * @memberof SyncDto
     */
    playlistName?: string;
    /**
     * 
     * @type {string}
     * @memberof SyncDto
     */
    youtubePlaylistId?: string;
    /**
     * 
     * @type {string}
     * @memberof SyncDto
     */
    spotifyPlaylistId?: string;
    /**
     * 
     * @type {string}
     * @memberof SyncDto
     */
    movesongEmail?: string;
    /**
     * 
     * @type {string}
     * @memberof SyncDto
     */
    date?: string;
    /**
     * 
     * @type {string}
     * @memberof SyncDto
     */
    lastSyncDate?: string;
    /**
     * 
     * @type {number}
     * @memberof SyncDto
     */
    interval?: number;
}